Extract RunnerPoller extensions and observers
Move large RunnerPoller responsibilities into focused extension files: add RunnerPoller+ApplyResult.swift (applyFetchResult, applyError and Sendable FetchError wrapper) and RunnerPoller+FetchAndEnrich.swift (fetchAndEnrichRunners). Also relocate the preference/scope observation starters into RunnerPollerObservers.swift and remove the duplicated implementations from RunnerPoller.swift. This refactor improves separation of concerns, reduces file size, and preserves behavior (including guarded Observable updates and Sendable-safe error bridging).
