+++ Sources/RunnerBarCore/Runner/Polling/RunnerPoller+ApplyResult.swift
@@ -1,0 +1,100 @@
+// RunnerPoller+ApplyResult.swift
+// RunnerBarCore
+
+import Foundation
+
+// swiftlint:disable:next missing_docs
+extension RunnerPoller {
+
+    // MARK: - Apply result
+
+    /// Merges a completed fetch into actor state and pushes the snapshot to `RunnerState`.
+    ///
+    /// Clears `state.fetchError` on every successful cycle so the UI error banner
+    /// dismisses automatically as soon as connectivity is restored. The write is
+    /// guarded — if `fetchError` is already `nil` the assignment is skipped to
+    /// avoid a spurious `@Observable` notification on every healthy poll cycle.
+    func applyFetchResult(
+        enrichedRunners: [Runner],
+        jobResult: JobPollResult,
+        groupResult: GroupPollResult
+    ) async {
+        let rateLimitSnapshot = await ghRateLimitSnapshot()
+        runners = enrichedRunners
+        jobs = jobResult.display
+        completedCache = jobResult.newCache
+        prevLiveJobs = jobResult.newPrevLive
+        actions = groupResult.display
+        actionGroupCache = groupResult.newGroupCache
+        prevLiveGroups = groupResult.newPrevLiveGroups
+        seenGroupIDs = groupResult.newSeenGroupIDs
+        isRateLimited = rateLimitSnapshot.isLimited
+        rateLimitResetDate = rateLimitSnapshot.resetDate
+        // swiftlint:disable:next line_length
+        log("RunnerPoller › fetch complete — actions=\(groupResult.display.count) jobs=\(jobResult.display.count) runners=\(enrichedRunners.count) isRateLimited=\(rateLimitSnapshot.isLimited) rateLimitResetDate=\(String(describing: rateLimitSnapshot.resetDate))", category: .runner)
+        await MainActor.run { [state] in
+            state.runners = enrichedRunners
+            state.jobs = jobResult.display
+            state.actions = groupResult.display
+            state.isRateLimited = rateLimitSnapshot.isLimited
+            state.rateLimitResetDate = rateLimitSnapshot.resetDate
+            if state.fetchError != nil { state.fetchError = nil }
+        }
+    }
+
+    /// Surfaces a fetch failure to the `RunnerState` read model.
+    ///
+    /// Mirrors `applyFetchResult` by updating both the actor-local rate-limit copies
+    /// (`self.isRateLimited`, `self.rateLimitResetDate` — read by `nextPollInterval()`)
+    /// and the `@Observable` read model (`state.*` — read by the view layer).
+    /// Without this sync, a failed cycle while rate-limited would leave the actor-local
+    /// copies stale, causing `nextPollInterval()` to compute the wrong cadence until the
+    /// next successful `applyFetchResult`.
+    ///
+    /// Snapshots rate-limit state so the UI never shows both banners simultaneously:
+    /// `clearGhRateLimit()` at the top of `fetchInternal()` clears the internal actor
+    /// before any throw, so this snapshot reflects the cleared state.
+    ///
+    /// The `fetchError` write is guarded by a `localizedDescription` comparison to avoid
+    /// re-notifying `@Observable` observers on every failed cycle when the message is
+    /// unchanged (e.g. sustained network loss).
+    ///
+    /// Intentionally does **not** clear `runners`, `jobs`, or `actions` — views show
+    /// stale data alongside the error banner rather than an empty list.
+    func applyError(_ error: any Error & Sendable) async {
+        let rateLimitSnapshot = await ghRateLimitSnapshot()
+        // Sync actor-local copies first — nextPollInterval() reads these directly.
+        isRateLimited = rateLimitSnapshot.isLimited
+        rateLimitResetDate = rateLimitSnapshot.resetDate
+        await MainActor.run { [state] in
+            // Guard the write: `any Error` is not Equatable, so compare via
+            // `localizedDescription` — the only field `fetchErrorBanner` consumes.
+            // Skipping the write when the message is unchanged avoids a spurious
+            // `@Observable` notification on every failed poll cycle.
+            if state.fetchError?.localizedDescription != error.localizedDescription {
+                state.fetchError = error
+            }
+            state.isRateLimited = rateLimitSnapshot.isLimited
+            state.rateLimitResetDate = rateLimitSnapshot.resetDate
+        }
+    }
+}
+
+// MARK: - FetchError
+
+// swiftlint:disable:next missing_docs
+extension RunnerPoller {
+
+    /// Sendable-safe wrapper that bridges an arbitrary `any Error` across an actor boundary.
+    ///
+    /// `any Error` is not `Sendable`, so passing it directly into `MainActor.run`
+    /// produces a warning under `-strict-concurrency=complete`. `FetchError` captures
+    /// `localizedDescription` — the only field read by `fetchErrorBanner` — and
+    /// re-surfaces it as a `LocalizedError` conformance so the message is preserved.
+    struct FetchError: LocalizedError, Sendable {
+        /// The user-facing description forwarded from the underlying error.
+        let errorDescription: String?
+        /// Wraps `underlying`, capturing its `localizedDescription` as a `Sendable` string.
+        init(_ underlying: any Error) { errorDescription = underlying.localizedDescription }
+    }
+}
