+++ Sources/RunnerBarCore/Runner/Polling/RunnerPoller.swift
@@ -130,73 +130,6 @@ public actor RunnerPoller {
         pollLoop.cancelAll()
     }
 
-    // MARK: - Observation loops
-
-    /// Starts (or restarts) the `pollingInterval` observation loop.
-    ///
-    /// Uses `AsyncStream<TimeInterval>` to match `PreferencesObserver.continuation` which is
-    /// typed `AsyncStream<TimeInterval>.Continuation` and yields
-    /// `TimeInterval(store.pollingInterval)`. The stream element type must match the
-    /// continuation type exactly — `pollingInterval` is an `Int` (seconds) but the observer
-    /// converts it to `TimeInterval` before yielding so the value can be used directly in
-    /// `nextPollInterval()` without a second conversion.
-    ///
-    /// **Self-cancellation avoidance**
-    /// `setIntervalObservationTask(newTask)` cancels the *previous* interval-observation
-    /// task and installs `newTask` as the new one. When called recursively from inside the
-    /// for-await body, the calling task must therefore create the new `Task` *before* passing
-    /// it to `setIntervalObservationTask` — otherwise the setter would cancel the caller
-    /// itself and the subsequent `start()` call would never execute.
-    private func startObservingPreferences() {
-        let injectedStore = preferencesStore
-        let newTask = Task { [weak self] in
-            let (stream, continuation) = AsyncStream<TimeInterval>.makeStream()
-            let observer: PreferencesObserver = await MainActor.run {
-                let preferencesObserver = PreferencesObserver(continuation: continuation, store: injectedStore)
-                preferencesObserver.start()
-                return preferencesObserver
-            }
-            for await newInterval in stream {
-                guard !Task.isCancelled else { break }
-                log("RunnerPoller › pollingInterval changed to \(Int(newInterval))s — restarting poll loop", category: .runner)
-                await self?.startObservingPreferences()
-                guard !Task.isCancelled else { break }
-                await self?.start()
-                break
-            }
-            _ = observer
-        }
-        pollLoop.setIntervalObservationTask(newTask)
-    }
-
-    /// Starts (or restarts) the `activeScopes` observation loop.
-    ///
-    /// **Self-cancellation avoidance**
-    /// Same pattern as `startObservingPreferences`: the new `Task` is created first,
-    /// then handed to `setScopeObservationTask` so the setter cancels the *previous*
-    /// task rather than the one currently executing.
-    private func startObservingScopes() {
-        let injectedStore = scopeStore
-        let newTask = Task { [weak self] in
-            let (stream, continuation) = AsyncStream<[String]>.makeStream()
-            let observer: ScopesObserver = await MainActor.run {
-                let scopesObserver = ScopesObserver(continuation: continuation, store: injectedStore)
-                scopesObserver.start()
-                return scopesObserver
-            }
-            for await _ in stream {
-                guard !Task.isCancelled else { break }
-                log("RunnerPoller › ScopeStore.activeScopes changed — restarting fetch", category: .runner)
-                await self?.startObservingScopes()
-                guard !Task.isCancelled else { break }
-                await self?.start()
-                break
-            }
-            _ = observer
-        }
-        pollLoop.setScopeObservationTask(newTask)
-    }
-
     // MARK: - Poll loop
 
     /// Starts (or restarts) the structured async poll loop.
@@ -400,209 +333,5 @@ public actor RunnerPoller {
             // Restore scope — not present in the API payload, must be carried forward.
             cache[cacheID] = updated.copying(scope: cached.scope)
         }
-    }
-
-    // MARK: - Apply result
-
-    /// Merges a completed fetch into actor state and pushes the snapshot to `RunnerState`.
-    ///
-    /// Clears `state.fetchError` on every successful cycle so the UI error banner
-    /// dismisses automatically as soon as connectivity is restored. The write is
-    /// guarded — if `fetchError` is already `nil` the assignment is skipped to
-    /// avoid a spurious `@Observable` notification on every healthy poll cycle.
-    private func applyFetchResult(
-        enrichedRunners: [Runner],
-        jobResult: JobPollResult,
-        groupResult: GroupPollResult
-    ) async {
-        let rateLimitSnapshot = await ghRateLimitSnapshot()
-        runners = enrichedRunners
-        jobs = jobResult.display
-        completedCache = jobResult.newCache
-        prevLiveJobs = jobResult.newPrevLive
-        actions = groupResult.display
-        actionGroupCache = groupResult.newGroupCache
-        prevLiveGroups = groupResult.newPrevLiveGroups
-        seenGroupIDs = groupResult.newSeenGroupIDs
-        isRateLimited = rateLimitSnapshot.isLimited
-        rateLimitResetDate = rateLimitSnapshot.resetDate
-        // swiftlint:disable:next line_length
-        log("RunnerPoller › fetch complete — actions=\(groupResult.display.count) jobs=\(jobResult.display.count) runners=\(enrichedRunners.count) isRateLimited=\(rateLimitSnapshot.isLimited) rateLimitResetDate=\(String(describing: rateLimitSnapshot.resetDate))", category: .runner)
-        await MainActor.run { [state] in
-            state.runners = enrichedRunners
-            state.jobs = jobResult.display
-            state.actions = groupResult.display
-            state.isRateLimited = rateLimitSnapshot.isLimited
-            state.rateLimitResetDate = rateLimitSnapshot.resetDate
-            if state.fetchError != nil { state.fetchError = nil }
-        }
-    }
-
-    /// Sendable-safe wrapper that bridges an arbitrary `any Error` across an actor boundary.
-    ///
-    /// `any Error` is not `Sendable`, so passing it directly into `MainActor.run`
-    /// produces a warning under `-strict-concurrency=complete`. `FetchError` captures
-    /// `localizedDescription` — the only field read by `fetchErrorBanner` — and
-    /// re-surfaces it as a `LocalizedError` conformance so the message is preserved.
-    private struct FetchError: LocalizedError, Sendable {
-        /// The user-facing description forwarded from the underlying error.
-        let errorDescription: String?
-        /// Wraps `underlying`, capturing its `localizedDescription` as a `Sendable` string.
-        init(_ underlying: any Error) { errorDescription = underlying.localizedDescription }
-    }
-
-    /// Surfaces a fetch failure to the `RunnerState` read model.
-    ///
-    /// Mirrors `applyFetchResult` by updating both the actor-local rate-limit copies
-    /// (`self.isRateLimited`, `self.rateLimitResetDate` — read by `nextPollInterval()`)
-    /// and the `@Observable` read model (`state.*` — read by the view layer).
-    /// Without this sync, a failed cycle while rate-limited would leave the actor-local
-    /// copies stale, causing `nextPollInterval()` to compute the wrong cadence until the
-    /// next successful `applyFetchResult`.
-    ///
-    /// Snapshots rate-limit state so the UI never shows both banners simultaneously:
-    /// `clearGhRateLimit()` at the top of `fetchInternal()` clears the internal actor
-    /// before any throw, so this snapshot reflects the cleared state.
-    ///
-    /// The `fetchError` write is guarded by a `localizedDescription` comparison to avoid
-    /// re-notifying `@Observable` observers on every failed cycle when the message is
-    /// unchanged (e.g. sustained network loss).
-    ///
-    /// Intentionally does **not** clear `runners`, `jobs`, or `actions` — views show
-    /// stale data alongside the error banner rather than an empty list.
-    private func applyError(_ error: any Error & Sendable) async {
-        let rateLimitSnapshot = await ghRateLimitSnapshot()
-        // Sync actor-local copies first — nextPollInterval() reads these directly.
-        isRateLimited = rateLimitSnapshot.isLimited
-        rateLimitResetDate = rateLimitSnapshot.resetDate
-        await MainActor.run { [state] in
-            // Guard the write: `any Error` is not Equatable, so compare via
-            // `localizedDescription` — the only field `fetchErrorBanner` consumes.
-            // Skipping the write when the message is unchanged avoids a spurious
-            // `@Observable` notification on every failed poll cycle.
-            if state.fetchError?.localizedDescription != error.localizedDescription {
-                state.fetchError = error
-            }
-            state.isRateLimited = rateLimitSnapshot.isLimited
-            state.rateLimitResetDate = rateLimitSnapshot.resetDate
-        }
-    }
-
-    // MARK: - fetchAndEnrichRunners
-
-    /// Fetches runners for the given scopes, resolves install paths, and enriches with metrics.
-    ///
-    /// `internal` — `fetch()` is the public entry point; this method is an implementation
-    /// detail not intended for direct external calls.
-    ///
-    /// **Phase 0** derives extra org scopes from local runners whose `gitHubUrl` points to a
-    /// single-path-component URL (org-only, not repo). This handles runners registered against
-    /// an org that the user hasn't explicitly added as a scope in ScopeStore — their org is
-    /// inferred from the local runner's URL so those runners continue to appear in the panel.
-    ///
-    /// **Phase 1** fans out concurrent scope fetches via `withTaskGroup`. Task completion order
-    /// is non-deterministic; views sort runners for display independently.
-    ///
-    /// **Phase 2** enriches each busy runner with system metrics concurrently.
-    ///
-    /// **Install-path lookup priority** (matches the original `RunnerStore`):
-    /// `byApiId ?? byAgentId ?? byFullKey ?? byName`
-    /// `byFullKey` ("scope/name" composite) ranks above `byName` so runners sharing
-    /// a name across different scopes resolve to the correct install path.
-    ///
-    /// - Parameters:
-    ///   - scopes: The active scopes to fetch runners for.
-    ///   - localRunners: The current local-runner snapshot (used for org-scope derivation).
-    ///   - installPathMap: Pre-built lookup maps from `buildInstallPathMap`.
-    internal func fetchAndEnrichRunners(
-        scopes: [String],
-        localRunners: [RunnerModel],
-        installPathMap: InstallPathMap
-    ) async -> [Runner] {
-        log("RunnerPoller › fetchAndEnrichRunners ENTER — scopes=\(scopes)", category: .runner)
-
-        // MARK: Phase 0 — Extra org-scope derivation from local runner URLs
-        // Delegates to `scopeFromUrl(_:)` in GitHubURLHelpers (F-52).
-        // Only org-scoped URLs produce a scope string without a "/"; repo-scoped
-        // URLs ("owner/repo") are filtered out by the `!contains("/")` guard below.
-        let configuredScopeSet = Set(scopes)
-        var extraOrgScopes: [String] = []
-        for localRunner in localRunners {
-            guard let url = localRunner.gitHubUrl,
-                  let derivedScope = scopeFromUrl(url),
-                  !derivedScope.contains("/") else { continue }
-            let orgScope = derivedScope
-            guard !configuredScopeSet.contains(orgScope),
-                  !extraOrgScopes.contains(orgScope)
-            else { continue }
-            extraOrgScopes.append(orgScope)
-            log("RunnerPoller › fetchAndEnrichRunners — derived extra org scope '\(orgScope)' from local runner '\(localRunner.runnerName)'", category: .runner)
-        }
-        if !extraOrgScopes.isEmpty {
-            log("RunnerPoller › fetchAndEnrichRunners — extra org scopes to fetch: \(extraOrgScopes)", category: .runner)
-        }
-
-        let allScopes = scopes + extraOrgScopes
-
-        // MARK: Phase 1 — Fetch raw runners for all scopes concurrently
-        var indexed: [IndexedScopedRunner] = []
-        await withTaskGroup(of: (String, [Runner]).self) { group in
-            for scope in allScopes {
-                group.addTask {
-                    let fetched = await fetchRunners(for: scope, decoder: self.decoder)
-                    return (scope, fetched)
-                }
-            }
-            for await (scope, fetched) in group {
-                indexed.append(contentsOf: fetched.map { IndexedScopedRunner(scope: scope, runner: $0) })
-            }
-        }
-
-        // MARK: Phase 2 — Enrich each busy runner with system metrics concurrently
-        // Lookup priority: byApiId ?? byAgentId ?? byFullKey ?? byName
-        let busyIndices = indexed.indices.filter { indexed[$0].runner.busy }
-        if !busyIndices.isEmpty {
-            let metricsResults: [(Int, RunnerMetrics?)] = await withTaskGroup(
-                of: (Int, RunnerMetrics?).self
-            ) { group in
-                for i in busyIndices {
-                    let runner = indexed[i].runner
-                    let scope = indexed[i].scope
-                    let installPath = installPathMap.byApiId[runner.id]
-                        ?? installPathMap.byAgentId[runner.id]
-                        ?? installPathMap.byFullKey["\(scope)/\(runner.name)"]
-                        ?? installPathMap.byName[runner.name]
-                    guard let path = installPath else {
-                        log("RunnerPoller › fetchAndEnrichRunners — no installPath for \(runner.name) id=\(runner.id) scope=\(scope)", category: .runner)
-                        continue
-                    }
-                    group.addTask {
-                        let metrics = await metricsForRunner(installPath: path)
-                        return (i, metrics)
-                    }
-                }
-                var results: [(Int, RunnerMetrics?)] = []
-                for await pair in group { results.append(pair) }
-                return results
-            }
-            for (i, metrics) in metricsResults {
-                indexed[i].runner = indexed[i].runner.copying(metrics: metrics)
-            }
-        }
-
-        let metricsUpdates = indexed.filter { $0.runner.busy && $0.runner.metrics != nil }
-        if !metricsUpdates.isEmpty {
-            for entry in metricsUpdates {
-#if DEBUG
-                // swiftlint:disable:next line_length
-                log("RunnerPoller › fetchAndEnrichRunners — applyMetrics: \(entry.runner.name) id=\(entry.runner.id) busy=\(entry.runner.busy) metrics=\(String(describing: entry.runner.metrics))", category: .runner)
-#endif
-                await applyMetrics(entry.runner.metrics, entry.runner.id, entry.runner.name)
-            }
-        }
-
-        let result = indexed.map(\.runner)
-        log("RunnerPoller › fetchAndEnrichRunners EXIT — returning \(result.count) runner(s)", category: .runner)
-        return result
     }
 }
